# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Krishna Pharmacy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: List prescriptions
      responses:
        "200":
          description: Prescriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: uploadPrescription
      tags: [prescriptions]
      summary: Upload a prescription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}/verify:
    patch:
      operationId: verifyPrescription
      tags: [prescriptions]
      summary: Approve or reject prescription (admin/pharmacist)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionVerify"
      responses:
        "200":
          description: Updated prescription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
...
components:
  schemas:
...
    Prescription:
      type: object
      required: [id, status, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: ["string", "null"]
        fileUrl:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PrescriptionInput:
      type: object
      required: [fileUrl, customerName]
      properties:
        fileUrl:
          type: string
        customerName:
          type: string
        notes:
          type: string
    PrescriptionVerify:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: string
...