# src/api-zod/generated/types/prescriptionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Krishna Pharmacy API
 * OpenAPI spec version: 0.1.0
 */

export interface PrescriptionInput {
  fileUrl: string;
  customerName: string;
  notes?: string;
}